file restructure, add ffi
